backend/requirements.txt
# Flask Core
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Authentication (NEW)
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
pyjwt==2.8.0

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# AI/ML
mistralai==0.0.11
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.2

# Vector Database
chromadb==0.4.18

# LangChain
langchain==0.1.0
langchain-community==0.0.10

# Utilities
requests==2.31.0
numpy==1.26.2
pandas==2.1.4